+++ R/golem.R
@@ -37,7 +37,7 @@
 #' @export
 scaffold_golem <- function(react = FALSE, vue = FALSE, use_cdn = TRUE, edit = interactive()){
   # checks
-  assert_that(has_npm())
+  assert_that(has_engine())
   assert_that(is_golem())
   assert_that(!has_scaffold(), msg = "Only a single golem scaffold is allowed")
   assert_that(!all(react, vue), msg = "Setup with either react or vue, not both")
@@ -45,13 +45,13 @@ scaffold_golem <- function(react = FALSE, vue = FALSE, use_cdn = TRUE, edit = in
   open_msg("golem")
 
   # init npm
-  npm_init()
+  engine_init()
 
   # install dependencies
   webpack_install()
 
   # edit package.json
-  npm_add_scripts()
+  engine_add_scripts()
 
   # set up dir for golem
   # only create dir if vue or react